+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-All notable changes to Mac Mouse Highlighter will be documented in this file.
+All notable changes to Mac Cursor Highlighter Thingy will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- Initial release of Mac Mouse Highlighter
+- Initial release of Mac Cursor Highlighter Thingy
 - **Highlight Styles**
   - Circle: Soft, filled circle with gradient effect
   - Spotlight: Screen dimming with spotlight around cursor